Rebase from Master

See merge request libracore/frappe!57

+++ frappe/utils/backups.py
@@ -189,7 +189,7 @@ def delete_temp_backups(older_than=24):
 		file_list = os.listdir(get_backup_path())
 		for this_file in file_list:
 			this_file_path = os.path.join(get_backup_path(), this_file)
-			if is_file_old(this_file_path, older_than):
+			if is_file_old(this_file_path, older_than) and os.path.isfile(this_file_path):
 				os.remove(this_file_path)
 
 def is_file_old(db_file_name, older_than=24):
